refactor(cli): own determinism command orchestration
