+++ Cargo.toml
@@ -94,7 +94,6 @@ copybook-corruption-predicates = { version = "=0.5.0", path = "crates/copybook-c
 copybook-dialect = { version = "=0.5.0", path = "crates/copybook-dialect" }
 copybook-codec = { version = "=0.5.0", path = "crates/copybook-codec" }
 copybook-codec-memory = { version = "=0.5.0", path = "crates/copybook-codec-memory" }
-copybook-cli-determinism = { version = "=0.5.0", path = "crates/copybook-cli-determinism" }
 copybook-sequence-ring = { version = "=0.5.0", path = "crates/copybook-sequence-ring" }
 copybook-arrow = { version = "=0.5.0", path = "crates/copybook-arrow" }
 copybook-governance = { version = "=0.5.0", path = "crates/copybook-governance" }
