+++ crates/copybook-cli-determinism/README.md
@@ -1,24 +1,21 @@
 # copybook-cli-determinism
 
-Determinism command execution for copybook CLI workflows.
+Retired 0.5 package for the determinism CLI implementation.
 
 ## Overview
 
-This crate implements the `determinism` subcommand for the copybook CLI. It parses CLI arguments
-and orchestrates decode, encode, and round-trip determinism checks, producing formatted output
-(text or JSON) that validates whether codec operations produce identical results across runs.
+The implementation is now owned directly by the binary-only `copybook-cli`
+package. The published 0.5 artifact remains available; no 0.6 release or new
+primary dependency is planned for this package.
 
 ## Usage
 
-```rust
-use copybook_cli_determinism::DeterminismCommand;
-use clap::Parser;
-
-// Typically used via the CLI:
-//   copybook determinism decode --format fixed --codepage cp037 schema.cpy data.bin
-//   copybook determinism encode --format fixed --codepage cp037 schema.cpy input.jsonl
-//   copybook determinism round-trip --format fixed --codepage cp037 schema.cpy data.bin
+```text
+copybook determinism decode --format fixed --codepage cp037 schema.cpy data.bin
 ```
+
+The command continues to support decode, encode, and round-trip validation
+through the CLI reference.
 
 ## License
