+++ crates/copybook-cli-determinism/claude.md
@@ -2,15 +2,12 @@
 # claude.md
 
 ## Scope
-- Determinism command execution and output rendering for the copybook CLI.
-- Layer: **CLI**.
-- Workspace package `copybook-cli-determinism` (see `Cargo.toml`).
+
+- Retired 0.5 package retained without the determinism implementation.
+- The live command owner is `copybook-cli`.
 
 ## Navigation
 - Crates index + layer map: `../claude.md`
 - Repository root + invariants: `../../CLAUDE.md`
 
 ## Build
-- `cargo build -p copybook-cli-determinism`
-- Test: `cargo test -p copybook-cli-determinism`
-- Lint: `cargo clippy -p copybook-cli-determinism -- -D warnings`
