+++ crates/copybook-cli-determinism/tests/claude.md
@@ -2,11 +2,10 @@
 # claude.md
 
 ## Scope
-- Integration tests for `copybook-cli-determinism` (CLI layer).
+
+- No 0.6 tests: command behavior is tested through `copybook-cli`.
 
 ## Navigation
+
 - Crate root: `../claude.md`
 - Repository root: `../../CLAUDE.md`
-
-## Run
-- `cargo test -p copybook-cli-determinism`
