+++ crates/copybook-cli/Cargo.toml
@@ -28,7 +28,6 @@ path = "src/main.rs"
 
 [dependencies]
 clap.workspace = true
-copybook-cli-determinism = { workspace = true }
 serde_json.workspace = true
 serde_plain.workspace = true
 tracing.workspace = true
